+++ contracts/ibc-reflect/tests/integration.rs
@@ -60,16 +60,13 @@ fn setup() -> Instance<MockApi, MockStorage, MockQuerier> {
 }
 
 fn fake_events(reflect_addr: &str) -> Vec<Event> {
-    let event = Event {
-        ty: "message".into(),
-        attributes: vec![
-            attr("module", "wasm"),
-            attr("signer", MOCK_CONTRACT_ADDR),
-            attr("code_id", "17"),
-            // We have to force this one to avoid the debug assertion against _
-            mock_wasmd_attr("_contract_address", reflect_addr),
-        ],
-    };
+    let event = Event::new("message").add_attributes(vec![
+        attr("module", "wasm"),
+        attr("signer", MOCK_CONTRACT_ADDR),
+        attr("code_id", "17"),
+        // We have to force this one to avoid the debug assertion against _
+        mock_wasmd_attr("_contract_address", reflect_addr),
+    ]);
     vec![event]
 }
 
@@ -91,7 +88,10 @@ fn connect<T: Into<String>>(
     let res: IbcBasicResponse = ibc_channel_connect(deps, mock_env(), handshake_connect).unwrap();
     assert_eq!(1, res.messages.len());
     assert_eq!(1, res.events.len());
-    assert_eq!(Event::new("ibc").attr("channel", "connect"), res.events[0]);
+    assert_eq!(
+        Event::new("ibc").add_attribute("channel", "connect"),
+        res.events[0]
+    );
     let id = res.messages[0].id;
 
     // fake a reply and ensure this works
@@ -227,7 +227,10 @@ fn handle_dispatch_packet() {
     // we didn't dispatch anything
     assert_eq!(0, res.messages.len());
     assert_eq!(1, res.events.len());
-    assert_eq!(Event::new("ibc").attr("packet", "receive"), res.events[0]);
+    assert_eq!(
+        Event::new("ibc").add_attribute("packet", "receive"),
+        res.events[0]
+    );
     // acknowledgement is an error
     let ack: AcknowledgementMsg<DispatchResponse> =
         from_slice(&res.acknowledgement, DESERIALIZATION_LIMIT).unwrap();