+++ contracts/reflect/tests/integration.rs
@@ -225,7 +225,7 @@ fn reply_and_query() {
 
     let id = 123u64;
     let data = Binary::from(b"foobar");
-    let events = vec![Event::new("message").attr("signer", "caller-addr")];
+    let events = vec![Event::new("message").add_attribute("signer", "caller-addr")];
     let result = ContractResult::Ok(SubMsgExecutionResponse {
         events: events.clone(),
         data: Some(data.clone()),